+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- 30 seconds timeout for next Ads
+- Allow enabling/disabling ads
+
 ## 0.1.0
 
 - Include a command line tool that allows you to update `AndroidManifest.xml` and `Info.plist` files.